Update README.md
I made this as a proof of concept and figured I could post the source code on git hub. It's made in python and uses a simple "code"(a = 1, b = 0, c = 1, etc) to encrypt a string inputted by the user. Feel free to use this is any way that you want to, I don't really care.